Move meteor-vite builder to build methods
